requirements.txt
telethon>=1.36.0
gspread>=6.1.0
google-auth>=2.30.0
python-dotenv>=1.0.1
tzdata>=2024.1
